Changes | Fixes | Stuff
misc_SpeedUpgrades.sqf - Checking INV_getItemTypeKg causes a script error that stops the entire blacklist check. Added AIR  vehicles to the blacklist.

_vars.sqf - Added back the speeding warrants that got overwritten.

startup_setVariables.sqf - Removed the OL_StatsLoadedFromDB waitUntil because it was missing a semicolon, along with setting the OL_StatsLoaded variable because it's set in a different file and has no reason to be checked / set here.

mission.sqm - Added a condition to all speedcams that checks if the player near the speedcam before calling the speedcam script (without this the speedcam gets ran for all clients when one client triggers it). Along with fixed the speedcam near sheriff town as it was trying to call the old DD function.